Functional websockets
